pyproject.toml
[project]
name = "brainwave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.1",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.4",
    "openai>=2.6.0",
    "pyyaml>=6.0.3",
    "scipy>=1.16.2",
    "uvicorn>=0.38.0",
    "websockets<14",
]
